MBS-11776 Extract Pod to RemoteDevice mapping
