+++ subprojects/test-runner/device-provider/impl/src/test/kotlin/com/avito/android/runner/devices/internal/kubernetes/KubernetesReservationClientTest.kt
@@ -41,7 +41,12 @@ internal class KubernetesReservationClientTest {
         dispatcher: CoroutineDispatcher = this.dispatcher,
     ): KubernetesReservationClient {
         return KubernetesReservationClient(
-            androidDebugBridge = androidDebugBridge,
+            deviceProvider = RemoteDeviceProviderImpl(
+                kubernetesApi,
+                StubEmulatorsLogsReporter,
+                androidDebugBridge,
+                StubLoggerFactory
+            ),
             kubernetesApi = kubernetesApi,
             emulatorsLogsReporter = StubEmulatorsLogsReporter,
             loggerFactory = StubLoggerFactory,
